# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AE Scientist API",
    "description": "Transform LLM conversations into actionable AE ideas",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "Get a paginated list of conversations for the current user.\n\nQuery Parameters:\n- conversation_status: Filter by \"draft\" or \"with_research\" (optional)\n- run_status: Filter by \"pending\", \"running\", \"completed\", or \"failed\" (optional)",
        "operationId": "list_conversations_api_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "conversation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Status"
            }
          },
          {
            "name": "run_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations/{conversation_id}/idea/research-run/{run_id}/events": {
      "get": {
        "tags": [
          "research-pipeline"
        ],
        "summary": "Stream Research Run Events",
        "operationId": "stream_research_run_events_api_conversations__conversation_id__idea_research_run__run_id__events_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Research pipeline progress events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunStreamEvent"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunStreamEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/idea/research-run/{run_id}/snapshot": {
      "get": {
        "tags": [
          "research-pipeline"
        ],
        "summary": "Get Research Run Snapshot",
        "operationId": "get_research_run_snapshot_api_conversations__conversation_id__idea_research_run__run_id__snapshot_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunInitialEventData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "List of conversations"
          }
        },
        "type": "object",
        "required": [
          "conversations"
        ],
        "title": "ConversationListResponse",
        "description": "Response for listing conversations."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchRunInitialEventData": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/ResearchRunInfo"
          },
          "stage_progress": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunStageProgress"
            },
            "type": "array",
            "title": "Stage Progress"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunLogEntry"
            },
            "type": "array",
            "title": "Logs"
          },
          "substage_events": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSubstageEvent"
            },
            "type": "array",
            "title": "Substage Events"
          },
          "substage_summaries": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSubstageSummary"
            },
            "type": "array",
            "title": "Substage Summaries"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunArtifactMetadata"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "tree_viz": {
            "items": {
              "$ref": "#/components/schemas/TreeVizItem"
            },
            "type": "array",
            "title": "Tree Viz"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "paper_generation_progress": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunPaperGenerationProgress"
            },
            "type": "array",
            "title": "Paper Generation Progress"
          },
          "best_node_selections": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunBestNodeSelection"
            },
            "type": "array",
            "title": "Best Node Selections"
          },
          "stage_skip_windows": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunStageSkipWindow"
            },
            "type": "array",
            "title": "Stage Skip Windows",
            "description": "Recorded windows when stages became skippable."
          },
          "hw_cost_estimate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchRunHwCostEstimateData"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hardware cost estimate available when the initial snapshot was emitted."
          },
          "hw_cost_actual": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchRunHwCostActualData"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hardware cost billed so far, if available."
          },
          "code_execution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchRunCodeExecution"
              },
              {
                "type": "null"
              }
            ],
            "description": "Latest code execution snapshot available for the run."
          }
        },
        "type": "object",
        "required": [
          "run",
          "stage_progress",
          "logs",
          "substage_events",
          "substage_summaries",
          "artifacts",
          "tree_viz",
          "events",
          "paper_generation_progress",
          "best_node_selections"
        ],
        "title": "ResearchRunInitialEventData"
      },
...
      "ResearchRunStreamEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ResearchRunInitialEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunStageProgressEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunRunEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunLogEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunArtifactEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunReviewCompletedEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunBestNodeEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunSubstageCompletedEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunPaperGenerationEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunSubstageEventStream"
          },
          {
            "$ref": "#/components/schemas/ResearchRunSubstageSummaryEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunCodeExecutionStartedEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunCodeExecutionCompletedEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunStageSkipWindowEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunHeartbeatEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunHwCostEstimateEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunHwCostActualEvent"
          },
          {
            "$ref": "#/components/schemas/ResearchRunErrorEvent"
          }
        ],
        "title": "ResearchRunStreamEvent",
        "description": "Root model for research pipeline SSE events.",
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "artifact": "#/components/schemas/ResearchRunArtifactEvent",
            "best_node_selection": "#/components/schemas/ResearchRunBestNodeEvent",
            "code_execution_completed": "#/components/schemas/ResearchRunCodeExecutionCompletedEvent",
            "code_execution_started": "#/components/schemas/ResearchRunCodeExecutionStartedEvent",
            "complete": "#/components/schemas/ResearchRunCompleteEvent",
            "error": "#/components/schemas/ResearchRunErrorEvent",
            "heartbeat": "#/components/schemas/ResearchRunHeartbeatEvent",
            "hw_cost_actual": "#/components/schemas/ResearchRunHwCostActualEvent",
            "hw_cost_estimate": "#/components/schemas/ResearchRunHwCostEstimateEvent",
            "initial": "#/components/schemas/ResearchRunInitialEvent",
            "log": "#/components/schemas/ResearchRunLogEvent",
            "paper_generation_progress": "#/components/schemas/ResearchRunPaperGenerationEvent",
            "review_completed": "#/components/schemas/ResearchRunReviewCompletedEvent",
            "run_event": "#/components/schemas/ResearchRunRunEvent",
            "stage_progress": "#/components/schemas/ResearchRunStageProgressEvent",
            "stage_skip_window": "#/components/schemas/ResearchRunStageSkipWindowEvent",
            "substage_completed": "#/components/schemas/ResearchRunSubstageCompletedEvent",
            "substage_event": "#/components/schemas/ResearchRunSubstageEventStream",
            "substage_summary": "#/components/schemas/ResearchRunSubstageSummaryEvent"
          }
        }
      },
...
    }
  }
}